pdf_table_extractor/requirements.txt
streamlit
pandas
camelot-py[cv]
openpyxl
pyyaml
langdetect
deep_translator
streamlit-authenticator>=0.2.2